# hello-goa/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Calculator Service
  description: HTTP service for adding numbers
  version: ""
host: localhost:80
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /multiply/{a}/{b}:
    get:
      tags:
      - hello-goa-service
      summary: multiply hello-goa-service
      operationId: hello-goa-service#multiply
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: integer
      - name: b
        in: path
        description: Right operand
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - grpc
      - http